[eugene.641] Spring Web MVC - 인증 관리 Step2 (#77)
* feat : 토큰 발급 api 작성

- [x] 인증 담당하는 컨트롤러 생성
- [x] 인증 및 토큰 발급 담당하는 서비스 생성

* feat : 토큰 인증 api 적용

- [x] 토큰을 인증하고 유저 정보를 가져오는 ArgumentResolver 생성
- [x] WebMvcConfigure 에 resolver 등록
- [x] 토큰을 이용하여 본인 정보 조회

* feat : 예약하기, 예약취소 개선 - 인증 적용

- [x] 예약 생성시 토큰 확인하여 유효한 경우에만 예약 가능하도록
- [x] 예약 삭제시 토큰 확인하여 유효한 경우에만 예약 취소하도록

* refactor : 예외처리 관련 예외 추가

- 예외처리 진행중

* feat: 테스트 코드 및 주석 추가

- 예약 권한 검사 테스트
- 테스트 코드 흐름 파악

* feat: 예약 권한 설정 및 예외 처리

- ResponseStatus 어노테이션을 사용한 예외처리로 변경
- 예약 생성 및 삭제 시 로그인 요구

* refactor: Response 클래스 추가

- controller에서 http응답으로 사용할 클래스 추가
- 기존 도메인 클래스에 변환 메소드 추가

+++ src/main/java/nextstep/reservation/ReservationService.java
@@ -1,30 +1,36 @@
 package nextstep.reservation;
 
+import nextstep.member.MemberDao;
 import nextstep.schedule.Schedule;
 import nextstep.schedule.ScheduleDao;
 import nextstep.support.DuplicateEntityException;
+import nextstep.support.NotExistEntityException;
 import nextstep.theme.Theme;
 import nextstep.theme.ThemeDao;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.stream.Collectors;
 
 @Service
 public class ReservationService {
-    public final ReservationDao reservationDao;
-    public final ThemeDao themeDao;
-    public final ScheduleDao scheduleDao;
+    private final ReservationDao reservationDao;
+    private final ThemeDao themeDao;
+    private final ScheduleDao scheduleDao;
+    private final MemberDao memberDao;
 
-    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao) {
+
+    public ReservationService(ReservationDao reservationDao, ThemeDao themeDao, ScheduleDao scheduleDao, MemberDao memberDao) {
         this.reservationDao = reservationDao;
         this.themeDao = themeDao;
         this.scheduleDao = scheduleDao;
+        this.memberDao = memberDao;
     }
 
-    public Long create(ReservationRequest reservationRequest) {
+    public Long create(ReservationRequest reservationRequest, Long memberId) {
         Schedule schedule = scheduleDao.findById(reservationRequest.getScheduleId());
         if (schedule == null) {
-            throw new NullPointerException();
+            throw new NotExistEntityException();
         }
 
         List<Reservation> reservation = reservationDao.findByScheduleId(schedule.getId());
@@ -34,27 +40,30 @@ public class ReservationService {
 
         Reservation newReservation = new Reservation(
                 schedule,
-                reservationRequest.getName()
+                memberDao.findById(memberId).getName()
         );
 
         return reservationDao.save(newReservation);
     }
 
-    public List<Reservation> findAllByThemeIdAndDate(Long themeId, String date) {
+    public List<ReservationResponse> findAllByThemeIdAndDate(Long themeId, String date) {
         Theme theme = themeDao.findById(themeId);
         if (theme == null) {
             throw new NullPointerException();
         }
 
-        return reservationDao.findAllByThemeIdAndDate(themeId, date);
+        List<Reservation> results = reservationDao.findAllByThemeIdAndDate(themeId, date);
+
+        return results.stream().map(e->e.toResponse()).collect(Collectors.toList());
     }
 
-    public void deleteById(Long id) {
+    public void deleteById(Long id, Long memberId) {
         Reservation reservation = reservationDao.findById(id);
         if (reservation == null) {
-            throw new NullPointerException();
+            throw new NotExistEntityException();
         }
-
-        reservationDao.deleteById(id);
+        if (reservation.getName().equals(memberDao.findById(memberId).getName())) {
+            reservationDao.deleteById(id);
+        }
     }
 }
